// apps/web/src/lib/graphql/__generated__/graphql.ts
export type AcfMediaItemConnection = Connection & MediaItemConnection & {
  __typename?: 'AcfMediaItemConnection';
  /** Edges for the AcfMediaItemConnection connection */
  edges: Array<AcfMediaItemConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<MediaItem>;
  /** Information about pagination in a connection. */
  pageInfo: AcfMediaItemConnectionPageInfo;
};
// ...
export type ContentSlices = AcfFieldGroup & AcfFieldGroupFields & ContentSlices_Fields & {
  __typename?: 'ContentSlices';
  /** Build the page/article body by adding content slices. */
  contentSlices?: Maybe<Array<Maybe<ContentSlicesContentSlices_Layout>>>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};
// ...
export type MediaDetails = {
  __typename?: 'MediaDetails';
  /** The filename of the mediaItem */
  file?: Maybe<Scalars['String']['output']>;
  /** The path to the mediaItem relative to the uploads directory */
  filePath?: Maybe<Scalars['String']['output']>;
  /** The height of the mediaItem */
  height?: Maybe<Scalars['Int']['output']>;
  /** Meta information associated with the mediaItem */
  meta?: Maybe<MediaItemMeta>;
  /** The available sizes of the mediaItem */
  sizes?: Maybe<Array<Maybe<MediaSize>>>;
  /** The width of the mediaItem */
  width?: Maybe<Scalars['Int']['output']>;
};
// ...
export type NodeWithFeaturedImageToMediaItemConnectionEdge = Edge & MediaItemConnectionEdge & OneToOneConnection & {
  __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The node of the connection, without the edges */
  node: MediaItem;
};
// ...
export type RootQuery = WithAcfOptionsPageHomepage & WithAcfOptionsPageSiteSettings & {
// ...
  pageBy?: Maybe<Page>;
// ...
};
// ...
export type GetPageByUriQueryVariables = Exact<{
  uri: Scalars['String']['input'];
}>;


export type GetPageByUriQuery = { __typename?: 'RootQuery', pageBy?: { __typename?: 'Page', databaseId: number, title?: string | null, slug?: string | null, date?: string | null, modified?: string | null, featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: { __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null, contentSlices?: { __typename?: 'ContentSlices', contentSlices?: Array<
        | { __typename?: 'ContentSlicesContentSlicesEmbedLayout', fieldGroupName?: string | null, url?: string | null }
        | { __typename?: 'ContentSlicesContentSlicesGalleryLayout', fieldGroupName?: string | null, images?: { __typename?: 'AcfMediaItemConnection', nodes: Array<{ __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null }> } | null }
        | { __typename?: 'ContentSlicesContentSlicesHeroLayout', fieldGroupName?: string | null, title?: string | null, subtitle?: string | null, ctaText?: string | null, ctaUrl?: string | null, image?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null }
        | { __typename?: 'ContentSlicesContentSlicesImageLayout', fieldGroupName?: string | null, caption?: string | null, altText?: string | null, image?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null }
        | { __typename?: 'ContentSlicesContentSlicesInfoboxLayout', fieldGroupName?: string | null, title?: string | null, content?: string | null }
        | { __typename?: 'ContentSlicesContentSlicesLatestPostsLayout', fieldGroupName?: string | null, title?: string | null, subtitle?: string | null, count?: number | null, category?: { __typename?: 'AcfTermNodeConnection', nodes: Array<
              | { __typename?: 'Category', slug?: string | null, databaseId: number }
              | { __typename?: 'PostFormat', databaseId: number }
              | { __typename?: 'Tag', databaseId: number }
            > } | null }
        | { __typename?: 'ContentSlicesContentSlicesQuoteLayout', fieldGroupName?: string | null, text?: string | null, author?: string | null }
        | { __typename?: 'ContentSlicesContentSlicesRichTextLayout', fieldGroupName?: string | null, content?: string | null }
       | null> | null } | null } | null };
// ...
export type GetPostBySlugQueryVariables = Exact<{
  slug: Scalars['String']['input'];
}>;


export type GetPostBySlugQuery = { __typename?: 'RootQuery', postBy?: { __typename?: 'Post', databaseId: number, title?: string | null, slug?: string | null, date?: string | null, modified?: string | null, excerpt?: string | null, featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: { __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null, categories?: { __typename?: 'PostToCategoryConnection', nodes: Array<{ __typename?: 'Category', databaseId: number, name?: string | null, slug?: string | null }> } | null, tags?: { __typename?: 'PostToTagConnection', nodes: Array<{ __typename?: 'Tag', databaseId: number, name?: string | null, slug?: string | null }> } | null, articleFields?: { __typename?: 'ArticleFields', subtitle?: string | null, featured?: boolean | null } | null, contentSlices?: { __typename?: 'ContentSlices', contentSlices?: Array<
        | { __typename?: 'ContentSlicesContentSlicesEmbedLayout', fieldGroupName?: string | null, url?: string | null }
        | { __typename?: 'ContentSlicesContentSlicesGalleryLayout', fieldGroupName?: string | null, images?: { __typename?: 'AcfMediaItemConnection', nodes: Array<{ __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null }> } | null }
        | { __typename?: 'ContentSlicesContentSlicesHeroLayout', fieldGroupName?: string | null, title?: string | null, subtitle?: string | null, ctaText?: string | null, ctaUrl?: string | null, image?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null }
        | { __typename?: 'ContentSlicesContentSlicesImageLayout', fieldGroupName?: string | null, caption?: string | null, altText?: string | null, image?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', databaseId: number, sourceUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null }
        | { __typename?: 'ContentSlicesContentSlicesInfoboxLayout', fieldGroupName?: string | null, title?: string | null, content?: string | null }
        | { __typename?: 'ContentSlicesContentSlicesLatestPostsLayout', fieldGroupName?: string | null, title?: string | null, subtitle?: string | null, count?: number | null, category?: { __typename?: 'AcfTermNodeConnection', nodes: Array<
              | { __typename?: 'Category', slug?: string | null, databaseId: number }
              | { __typename?: 'PostFormat', databaseId: number }
              | { __typename?: 'Tag', databaseId: number }
            > } | null }
        | { __typename?: 'ContentSlicesContentSlicesQuoteLayout', fieldGroupName?: string | null, text?: string | null, author?: string | null }
        | { __typename?: 'ContentSlicesContentSlicesRichTextLayout', fieldGroupName?: string | null, content?: string | null }
       | null> | null } | null } | null };
// ...
export const GetPageByUriDocument = new TypedDocumentString(`
    query GetPageByUri($uri: String!) {
  pageBy(uri: $uri) {
    ...PageDetailFields
  }
}
    fragment MediaItemFields on MediaItem {
  databaseId
  sourceUrl
  altText
  mediaDetails {
    width
    height
  }
}
fragment PageDetailFields on Page {
  databaseId
  title
  slug
  date
  modified
  featuredImage {
    node {
      ...MediaItemFields
    }
  }
  contentSlices {
    contentSlices {
      ...SliceFields
    }
  }
}
fragment SliceFields on ContentSlicesContentSlices_Layout {
  ... on ContentSlicesContentSlicesRichTextLayout {
    fieldGroupName
    content
  }
  ... on ContentSlicesContentSlicesImageLayout {
    fieldGroupName
    image {
      node {
        ...MediaItemFields
      }
    }
    caption
    altText
  }
  ... on ContentSlicesContentSlicesGalleryLayout {
    fieldGroupName
    images {
      nodes {
        ...MediaItemFields
      }
    }
  }
  ... on ContentSlicesContentSlicesQuoteLayout {
    fieldGroupName
    text
    author
  }
  ... on ContentSlicesContentSlicesInfoboxLayout {
    fieldGroupName
    title
    content
  }
  ... on ContentSlicesContentSlicesHeroLayout {
    fieldGroupName
    title
    subtitle
    image {
      node {
        ...MediaItemFields
      }
    }
    ctaText
    ctaUrl
  }
  ... on ContentSlicesContentSlicesLatestPostsLayout {
    fieldGroupName
    title
    subtitle
    category {
      nodes {
        databaseId
        ... on Category {
          slug
        }
      }
    }
    count
  }
  ... on ContentSlicesContentSlicesEmbedLayout {
    fieldGroupName
    url
  }
}`) as unknown as TypedDocumentString<GetPageByUriQuery, GetPageByUriQueryVariables>;
// ...
export const GetPostBySlugDocument = new TypedDocumentString(`
    query GetPostBySlug($slug: String!) {
  postBy(slug: $slug) {
    ...PostDetailFields
  }
}
    fragment CategoryFields on Category {
  databaseId
  name
  slug
}
fragment MediaItemFields on MediaItem {
  databaseId
  sourceUrl
  altText
  mediaDetails {
    width
    height
  }
}
fragment PostDetailFields on Post {
  databaseId
  title
  slug
  date
  modified
  excerpt
  featuredImage {
    node {
      ...MediaItemFields
    }
  }
  categories {
    nodes {
      ...CategoryFields
    }
  }
  tags {
    nodes {
      ...TagFields
    }
  }
  articleFields {
    subtitle
    featured
  }
  contentSlices {
    contentSlices {
      ...SliceFields
    }
  }
}
fragment SliceFields on ContentSlicesContentSlices_Layout {
  ... on ContentSlicesContentSlicesRichTextLayout {
    fieldGroupName
    content
  }
  ... on ContentSlicesContentSlicesImageLayout {
    fieldGroupName
    image {
      node {
        ...MediaItemFields
      }
    }
    caption
    altText
  }
  ... on ContentSlicesContentSlicesGalleryLayout {
    fieldGroupName
    images {
      nodes {
        ...MediaItemFields
      }
    }
  }
  ... on ContentSlicesContentSlicesQuoteLayout {
    fieldGroupName
    text
    author
  }
  ... on ContentSlicesContentSlicesInfoboxLayout {
    fieldGroupName
    title
    content
  }
  ... on ContentSlicesContentSlicesHeroLayout {
    fieldGroupName
    title
    subtitle
    image {
      node {
        ...MediaItemFields
      }
    }
    ctaText
    ctaUrl
  }
  ... on ContentSlicesContentSlicesLatestPostsLayout {
    fieldGroupName
    title
    subtitle
    category {
      nodes {
        databaseId
        ... on Category {
          slug
        }
      }
    }
    count
  }
  ... on ContentSlicesContentSlicesEmbedLayout {
    fieldGroupName
    url
  }
}
fragment TagFields on Tag {
  databaseId
  name
  slug
}`) as unknown as TypedDocumentString<GetPostBySlugQuery, GetPostBySlugQueryVariables>;